rename swift promisekit test util test case & fix expected string
and fix xome silly typos. Damn you Xcode!!!
